Extracts Knot.x server module, removes knotx core module
