+++ launcher/src/main/resources/META-INF/services/io.vertx.config.spi.ConfigProcessor
