+++ server-http/core/LICENSE
