+++ server-http/core/build.gradle.kts
@@ -30,18 +30,14 @@ group = "io.knotx"
 // Dependencies
 // -----------------------------------------------------------------------------
 
-apply(from = "../gradle/common.deps.gradle.kts")
-apply(from = "../gradle/codegen.deps.gradle.kts")
+apply(from = "../../gradle/common.deps.gradle.kts")
+apply(from = "../../gradle/codegen.deps.gradle.kts")
 dependencies {
-  api(project(":knotx-fragment-api"))
   api(project(":knotx-server-http-api"))
-  api(project(":knotx-knot-engine-api"))
   api(group = "ch.qos.logback", name = "logback-classic")
   api(group = "com.google.guava", name = "guava")
   api(group = "commons-io", name = "commons-io")
   api(group = "org.apache.commons", name = "commons-lang3")
-  api(group = "org.jsoup", name = "jsoup")
-  api(group = "com.typesafe", name = "config")
   api(group = "commons-collections", name = "commons-collections")
 }
 
@@ -49,12 +45,9 @@ dependencies {
 // Source sets
 // -----------------------------------------------------------------------------
 
-apply(from = "../gradle/common.gradle.kts")
+apply(from = "../../gradle/common.gradle.kts")
 sourceSets.named("main") {
   java.srcDir("src/main/generated")
-}
-sourceSets.named("test") {
-  java.srcDirs(listOf("src/main/java", "src/main/generated"))
 }
 
 // -----------------------------------------------------------------------------
@@ -92,13 +85,13 @@ tasks.register<Jar>("javadocJar") {
 publishing {
   publications {
     create<MavenPublication>("mavenJava") {
-      artifactId = "knotx-core"
+      artifactId = "knotx-server-http-core"
       from(components["java"])
       artifact(tasks["sourcesJar"])
       artifact(tasks["javadocJar"])
       pom {
-        name.set("Knot.x Core")
-        description.set("Knot.x - efficient, high-performance and scalable integration platform for modern websites")
+        name.set("Knot.x Core HTTP Server")
+        description.set("Knot.x HTTP Server that handles all incoming HTTP requests.")
         url.set("http://knotx.io")
         licenses {
           license {
