+++ server-http/core/src/main/asciidoc/dataobjects.adoc
