+++ server-http/core/src/main/java/io/knotx/server/configuration/KnotxServerOptions.java
