+++ server-http/core/src/main/java/io/knotx/server/handler/failure/ErrorRoutingHandlerFactory.java
