+++ server-http/core/src/main/java/io/knotx/server/handler/http/response/header/CustomHeaderRoutingHandlerFactory.java
