+++ server-http/core/src/main/resources/io/knotx/logging/logback/console-appender.xml
