+++ server-http/core/src/main/resources/io/knotx/logging/logback/defaults.xml
