+++ server-http/core/src/main/resources/knotx-codegen/common-lib.templ
