+++ server-http/core/src/main/resources/knotx-codegen/dataObjectCheatsheet.templ
